Warning 5

+++ distributedcombigrid/src/sgpp/distributedcombigrid/manager/ProcessManager.hpp
@@ -319,7 +319,7 @@ inline void ProcessManager::recomputeOptimumCoefficients(std::string prob_name,
       std::cout << newCoeffs[i] << " ";
     }
     std::cout << "\n";
-    int roundedSum = round(sum);
+    int roundedSum = boost::numeric_cast<int>(round(sum));
     std::cout << "Coefficient sum: " << roundedSum << "\n";
 
     assert(roundedSum == 1);
